Fix postprocessing for PrusaSlicer 2.4.0.
Postprocessing will now be done in a temporary location and then the gcode file is moved to the final location. This final path will be in an environment variable `SLIC3R_PP_OUTPUT_NAME`, so use that to determine the paths of any additional output files besides the gcode file.