Separate the configuration from the script.
Requiring users to edit the Perl script is a bad idea, and it made
updates annoying because the configuration had to be manually
transferred to new editions of the script.

The configuration is now in a file that by default is looked for in the
same path as the script. Other locations can be passed with `-f`.

You will have to do a one-time migration to this new config system.
Migration tips:
* Convert single to double quotes and omit commas in lists.
* Do not copy the semicolons.